@@ -18,38 +18,31 @@
                     OnVisibleChange="@OnOverlayTriggerVisibleChange">
         <Overlay>
             <CascadingValue Value="this" IsFixed="@true">
-                @if (ShowPanel && !_isOptionsZero)
-                {
-                    <div style="max-height: 256px; overflow-y: auto; overflow-anchor: none; @_minWidth">
-                        <div style="display: flex;flex-direction: column;">
-
-                            @if (OverlayTemplate != null)
-                            {
-                                @OverlayTemplate
-                            }
-                            else
+                <div style="max-height: 256px; overflow-y: auto; overflow-anchor: none; @_minWidth">
+                    <div style="display: flex;flex-direction: column;">
+                        @if (OverlayTemplate != null)
+                        {
+                            @OverlayTemplate
+                        }
+                        else if (_filteredOptions?.Any()==true)
+                        {
+                            @foreach (var option in _filteredOptions)
                             {
                                 @if (OptionTemplate != null)
                                 {
-                                    @foreach (var option in GetOptionItems())
-                                    {
-                                        @OptionTemplate(option)
-                                    }
+                                    @OptionTemplate(option)
                                 }
                                 else
                                 {
-                                    @foreach (var option in GetOptionItems())
-                                    {
-                                        var label = OptionFormat == null ? option.Label : OptionFormat(option);
-                                        <AutoCompleteOption Value="@option.Value" Label="@option.Label" Disabled="option.IsDisabled">
-                                            @label
-                                        </AutoCompleteOption>
-                                    }
+                                    var label = OptionFormat == null ? option.Label : OptionFormat(option);
+                                    <AutoCompleteOption Value="@option.Value" Label="@option.Label" Disabled="option.IsDisabled">
+                                        @label
+                                    </AutoCompleteOption>
                                 }
                             }
-                        </div>
+                        }
                     </div>
-                }
+                </div>
             </CascadingValue>
         </Overlay>
         <Unbound>
