@@ -755,13 +755,13 @@ namespace AntDesign
                 return;
             }
 
-            if (string.IsNullOrWhiteSpace(_searchValue))
-            {
-                return;
-            }
+            _selectContent?.ClearSearch();
 
-            _searchValue = string.Empty;
-            _prevSearchValue = string.Empty;
+            if (!string.IsNullOrWhiteSpace(_searchValue))
+            {
+                _searchValue = string.Empty;
+                _prevSearchValue = string.Empty;
+            }
 
             if (SelectMode != SelectMode.Default && HideSelected)
             {
@@ -779,6 +779,11 @@ namespace AntDesign
                 SelectOptionItems.Where(x => x.IsHidden)
                     .ForEach(i => i.IsHidden = false);
             }
+        }
+
+        protected void OnOverlayShow()
+        {
+            _selectContent?.DiscoverySearch();
         }
 
         /// <summary>
@@ -1102,7 +1107,6 @@ namespace AntDesign
         internal async Task CloseAsync()
         {
             await _dropDown.Hide(true);
-            _selectContent?.ClearSearch();
         }
 
         /// <summary>
