@@ -14,7 +14,7 @@
                 {
                     <div class="ant-steps-progress-icon">
                         <Progress Percent="@((double) Percent)"
-                                  Width=@(Size == "small" ? 32 : 40)
+                                  Width=@(Size == StepsSize.Small ? 32 : 40)
                                   Type=ProgressType.Circle
                                   StrokeWidth="4"
                                   Format=@((p) => "") />
@@ -22,22 +22,22 @@
 
                             @if (string.IsNullOrEmpty(Icon))
                             {
-                                if (Status == "finish")
+                                if (Status == StepsStatus.Finish)
                                 {
                                     <Icon Type="check"></Icon>
                                 }
-                                else if (Status == "error")
+                                else if (Status == StepsStatus.Error)
                                 {
                                     <Icon Type="close"></Icon>
                                 }
-                                else if (Status == "process" || Status == "wait")
+                                else if (Status == StepsStatus.Process || Status == StepsStatus.Wait)
                                 {
                                     @(Index + 1)
                                 }
                             }
                             else
                             {
-                                if (Status == "error")
+                                if (Status == StepsStatus.Error)
                                 {
                                     <Icon Type="close"></Icon>
                                 }
@@ -55,22 +55,22 @@
 
                         @if (string.IsNullOrEmpty(Icon))
                         {
-                            if (Status == "finish")
+                            if (Status == StepsStatus.Finish)
                             {
                                 <Icon Type="check"></Icon>
                             }
-                            else if (Status == "error")
+                            else if (Status == StepsStatus.Error)
                             {
                                 <Icon Type="close"></Icon>
                             }
-                            else if (Status == "process" || Status == "wait")
+                            else if (Status == StepsStatus.Process || Status == StepsStatus.Wait)
                             {
                                 @(Index + 1)
                             }
                         }
                         else
                         {
-                            if (Status == "error")
+                            if (Status == StepsStatus.Error)
                             {
                                 <Icon Type="close"></Icon>
                             }
