@@ -83,10 +83,11 @@
              role="tabpanel"
              aria-hidden="@(_isActive ? "false" : "true")"
              aria-labelledby="@($"rc-tabs-{Id}-tab-{Key}")">
-            @if (_isActive || _hasRendered || ForceRender)
+            @if (ChildContent != null && (_isActive || _hasRendered || ForceRender))
             {
                 _hasRendered = true;
-                @ChildContent
+                <AntDesign.Core.Internal.KeepAlive ChildContent="ChildContent" IsActive="@_isActive"></AntDesign.Core.Internal.KeepAlive>
+                
             }
         </div>
     }
