@@ -1,5 +1,5 @@
 <div>
-    <Steps Current=current OnChange=onChange>
+    <Steps Current="current" OnChange="onChange">
         <Step Title="Step 1" Description="This is a description." />
         <Step Title="Step 2" Description="This is a description." />
         <Step Title="Step 3" Description="This is a description." />
@@ -7,7 +7,7 @@
 
     <Divider />
 
-    <Steps Current=current OnChange=onChange Direction="vertical">
+    <Steps Current="current" OnChange="onChange" Direction="@StepsDirection.Vertical">
         <Step Title="Step 1" Description="This is a description." />
         <Step Title="Step 2" Description="This is a description." />
         <Step Title="Step 3" Description="This is a description." />
