@@ -1,4 +1,4 @@
-<Steps Current=1 ProgressDot=@customDot>
+<Steps Current="1" ProgressDot="@customDot">
     <Step Title="Finished" Description="You can hover on the dot." />
     <Step Title="In Progress" Description="You can hover on the dot." />
     <Step Title="Waiting" Description="You can hover on the dot." />
