@@ -1,5 +1,5 @@
 <div>
-    <Steps Current="1" Status="error">
+    <Steps Current="1" Status="@StepsStatus.Error">
         <Step Title="Finished" Description="This is a description." />
         <Step Title="In Progress" Description="This is a description." />
         <Step Title="Waiting" Description="This is a description." />
