@@ -1,8 +1,8 @@
 <div>
     <Steps>
-        <Step Status="finish" Title="Login" Icon="user" />
-        <Step Status="finish" Title="Verification" Icon="solution"  />
-        <Step Status="process" Title="Pay" Icon="loading" />
-        <Step Status="wait" Title="Done" Icon="smile"/>
+        <Step Status="@StepsStatus.Finish" Title="Login" Icon="user" />
+        <Step Status="@StepsStatus.Finish" Title="Verification" Icon="solution"  />
+        <Step Status="@StepsStatus.Process" Title="Pay" Icon="loading" />
+        <Step Status="@StepsStatus.Wait" Title="Done" Icon="smile"/>
     </Steps>
 </div>
