@@ -5,7 +5,7 @@
         <Step Title="Waiting" Description="This is a description." />
     </Steps>
     <Divider />
-    <Steps ShowProgressDot Current="1" Direction="vertical">
+    <Steps ShowProgressDot Current="1" Direction="@StepsDirection.Vertical">
         <Step Title="Finished" Description="This is a description. This is a description." />
         <Step Title="Finished" Description="This is a description. This is a description." />
         <Step Title="In Progress" Description="This is a description. This is a description." />
