@@ -1,5 +1,5 @@
 <div>
-    <Steps Size="small" Current=1>
+    <Steps Size="@StepsSize.Small" Current=1>
         <Step Title="Finished" />
         <Step Title="In Progress" />
         <Step Title="Waiting" />
