@@ -1,5 +1,5 @@
 <div>
-    <Steps Direction="vertical" Current="1">
+    <Steps Direction="@StepsDirection.Vertical" Current="1">
         <Step Title="Finished" Description="This is a description." />
         <Step Title="In Progress" Description="This is a description." />
         <Step Title="Waiting" Description="This is a description." />
